refactor: CupRanks - CupRankDto List 변환 생성자 삭제 - 도메인 CupRankDto vo 삭제 및 UpdateCupRankDto -> CupRankDto 이름 변경
