--- backend/src/main/java/backend/mulkkam/cup/domain/dto/CupRankDto.java
@@ -1,7 +1,0 @@
-package backend.mulkkam.cup.domain.dto;
-
-public record CupRankDto(
-        Long id,
-        Integer rank
-) {
-}
