+++ backend/src/main/java/backend/mulkkam/cup/dto/CupRankDto.java
@@ -1,0 +1,13 @@
+package backend.mulkkam.cup.dto;
+
+import backend.mulkkam.cup.domain.Cup;
+
+public record CupRankDto(
+        Long id,
+        Integer rank
+) {
+
+    public CupRankDto(Cup cup) {
+        this(cup.getId(), cup.getCupRank().value());
+    }
+}
