+++ backend/src/main/java/backend/mulkkam/cup/dto/request/UpdateCupRanksRequest.java
@@ -1,8 +1,8 @@
 package backend.mulkkam.cup.dto.request;
 
-import backend.mulkkam.cup.dto.UpdateCupRankDto;
+import backend.mulkkam.cup.dto.CupRankDto;
 
 import java.util.List;
 
-public record UpdateCupRanksRequest(List<UpdateCupRankDto> cups) {
+public record UpdateCupRanksRequest(List<CupRankDto> cups) {
 }
