+++ backend/src/main/java/backend/mulkkam/cup/dto/response/CupsRanksResponse.java
@@ -1,8 +1,8 @@
 package backend.mulkkam.cup.dto.response;
 
-import backend.mulkkam.cup.dto.UpdateCupRankDto;
+import backend.mulkkam.cup.dto.CupRankDto;
 
 import java.util.List;
 
-public record CupsRanksResponse(List<UpdateCupRankDto> cups) {
+public record CupsRanksResponse(List<CupRankDto> cups) {
 }
