+++ backend/src/main/java/backend/mulkkam/cup/service/CupService.java
@@ -2,32 +2,33 @@ package backend.mulkkam.cup.service;
 
 import backend.mulkkam.common.exception.CommonException;
 import backend.mulkkam.cup.domain.Cup;
+import backend.mulkkam.cup.domain.IntakeType;
 import backend.mulkkam.cup.domain.collection.CupRanks;
-import backend.mulkkam.cup.domain.dto.CupRankDto;
 import backend.mulkkam.cup.domain.vo.CupAmount;
 import backend.mulkkam.cup.domain.vo.CupNickname;
 import backend.mulkkam.cup.domain.vo.CupRank;
-import backend.mulkkam.cup.dto.UpdateCupRankDto;
+import backend.mulkkam.cup.dto.CupRankDto;
 import backend.mulkkam.cup.dto.request.CupNicknameAndAmountModifyRequest;
 import backend.mulkkam.cup.dto.request.CupRegisterRequest;
 import backend.mulkkam.cup.dto.request.UpdateCupRanksRequest;
 import backend.mulkkam.cup.dto.response.CupResponse;
 import backend.mulkkam.cup.dto.response.CupsRanksResponse;
-import backend.mulkkam.cup.domain.IntakeType;
 import backend.mulkkam.cup.dto.response.CupsResponse;
 import backend.mulkkam.cup.repository.CupRepository;
 import backend.mulkkam.member.domain.Member;
 import backend.mulkkam.member.repository.MemberRepository;
-
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
+
 import static backend.mulkkam.common.exception.errorCode.BadRequestErrorCode.INVALID_CUP_COUNT;
+import static backend.mulkkam.common.exception.errorCode.ConflictErrorCode.DUPLICATED_CUP;
 import static backend.mulkkam.common.exception.errorCode.ForbiddenErrorCode.NOT_PERMITTED_FOR_CUP;
 import static backend.mulkkam.common.exception.errorCode.NotFoundErrorCode.NOT_FOUND_CUP;
 import static backend.mulkkam.common.exception.errorCode.NotFoundErrorCode.NOT_FOUND_MEMBER;
@@ -71,11 +72,7 @@ public class CupService {
             UpdateCupRanksRequest request,
             Long memberId
     ) {
-        List<CupRankDto> cupRankDtos = request.cups()
-                .stream()
-                .map(UpdateCupRankDto::toCupRankDto)
-                .toList();
-        CupRanks cupRanks = new CupRanks(cupRankDtos);
+        CupRanks cupRanks = new CupRanks(buildCupRankMapById(request.cups()));
         List<Cup> cups = getAllByIdsAndMemberId(cupRanks.getCupIds(), memberId);
 
         for (Cup cup : cups) {
@@ -84,9 +81,20 @@ public class CupService {
 
         return new CupsRanksResponse(
                 cups.stream()
-                        .map(UpdateCupRankDto::new)
+                        .map(CupRankDto::new)
                         .toList()
         );
+    }
+
+    private Map<Long, CupRank> buildCupRankMapById(List<CupRankDto> cupRanks) {
+        Map<Long, CupRank> ranks = new HashMap<>();
+        for (CupRankDto cup : cupRanks) {
+            if (ranks.containsKey(cup.id())) {
+                throw new CommonException(DUPLICATED_CUP);
+            }
+            ranks.put(cup.id(), new CupRank(cup.rank()));
+        }
+        return ranks;
     }
 
     private List<Cup> getAllByIdsAndMemberId(
